requirements.txt
streamlit
pydantic
python-dotenv
openai
langchain
chromadb
PyPDF4
pdfplumber
tiktoken
watchdog